[WM] send user command msgs outside go blocks
This means the message will be on its way already while the go block is
parking the thread to wait for the response to come down the response
channel. The way it was before we were making the go thread kick off the
message send and then re-park to wait for the response. H/T to Eric for
the suggestion.