[WM] move users resource to root URL path
I realized when writing the "put this into synapse-balancer" card that
we actually wanted "users" to be the root resource here. Otherwise all
of the external API URLs would end up looking like
https://api.turbovote.org/users/users/... because the first "/users" is
what routes it to this component and the second would then be needed to
match the old routes in here. However, since this is a "user" API, it
makes sense for that to be the root resource and thus our external API
URLs will be like this: https://api.turbovote.org/users/...