[WM] update pedestal-toolbox to 0.6.2
Fixes a problem with Content-Type regexes where we weren't matching some valid ones when they e.g. had a charset appended to the MIME type string. Firefox likes to send them this way, and we should be handling them correctly regardless. H/T to Chris for the fix in pedestal-toolbox.